Fixing a bunch of the issues keeping clang-tidy unhappy

+++ src/boot/osloader/graphics.c
@@ -97,7 +97,7 @@ EFI_STATUS OslpInitializeGraphics(
     *FramebufferHeight = Gop->Mode->Info->VerticalResolution;
     *FramebufferPitch = Gop->Mode->Info->PixelsPerScanLine * 4;
 
-    uint64_t FrameBufferSize = *FramebufferHeight * *FramebufferPitch;
+    uint64_t FrameBufferSize = (uint64_t)(*FramebufferHeight * *FramebufferPitch);
     uint64_t FrontBufferSize = (FrameBufferSize + SIZE_2MB - 1) & ~(SIZE_2MB - 1);
     *FrontBufferPhys = OslAllocatePages(FrontBufferSize, SIZE_2MB);
     if (!*FrontBufferPhys) {
